Move Input from client to presto-main
It's not needed by the client (yet). We may move it back later if we decide to expose
this info to clients in a structural way
