+++ presto-main/src/test/java/com/facebook/presto/execution/TestInput.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.client;
+package com.facebook.presto.execution;
 
 import com.google.common.collect.ImmutableList;
 import io.airlift.json.JsonCodec;
